different levels of normalization
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@2327 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ simcraft.h
@@ -353,8 +353,8 @@ struct gear_stats_t
 // normalized and standard random
 struct roll_instance_t
 {
-	int    given;
-	double wanted; // sum of chances
+	double given;
+	double wanted; 
 	int    nTries;
 	roll_instance_t() { memset( (void*) this, 0x00, sizeof( roll_instance_t ) ); }
 };
